.history/requirements_20250615193856.txt
Flask==3.0.0
Werkzeug==3.0.0
python-dotenv==1.0.0
cloudinary==1.36.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
gunicorn==21.2.0
flask-session==0.5.0
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3  # Better parser for BeautifulSoup
Pillow==10.0.0
google-generativeai==0.3.2
oauthlib==3.2.2
requests-oauthlib==1.3.1
tweepy==4.14.0
urllib3==2.0.7
certifi==2023.11.17
# Added for stability
Flask-Caching==2.1.0
pytz==2023.3
APScheduler==3.10.4